refactor(rust-api): extract training routes + helpers into training.rs (sc-1636)
Move the training cluster out of lib.rs into training.rs: the dataset
CRUD handlers, the caption-sidecar + caption-job handlers,
create_training_job, and the training-specific helpers (base-model
resolution, output-scope/location, disk-space checks, adapter-file
trust). get_project_file stays at the root (not training).

Same pattern: `use super::*;` + pub(crate) on the cluster. The whole
cluster is pub(crate) because trusted_adapter_files is also called from
the LoRA-registration code outside the module; tests.rs's
insufficient_disk_space import is repointed to super::training.
Behavior-preserving. lib.rs 7,475 -> 6,797 lines.
